Revert "* fix researchDurationDivisor"
This reverts commit 0183fc97569f88cfeb813f903cebfc6b51ffc242.

+++ pkg/wrapper/ogame.go
@@ -4602,7 +4602,7 @@ func (b *OGame) GetUsername() string {
 }
 
 // GetResearchSpeed gets the research speed
-func (b *OGame) GetResearchSpeed() float64 {
+func (b *OGame) GetResearchSpeed() int64 {
 	return b.serverData.ResearchDurationDivisor
 }
